pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "container-defense-stack"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "pydantic>=2.0", "click>=8.0", "rich>=13.0", "jsonschema>=4.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
addopts = "--cov=validators --cov-report=term-missing --cov-fail-under=70"
